차트에 소수점이 추가가 안되는 오류 수정

+++ MuscleDiary/app/src/main/java/com/munchkin/musclediary/src/main/chart/models/WeightResult.java
@@ -4,12 +4,12 @@ import com.google.gson.annotations.SerializedName;
 
 public class WeightResult {
     @SerializedName("weight")
-    private int weight;
+    private float weight;
 
     @SerializedName("recordDate")
     private String recordDate;
 
-    public int getWeight() {
+    public float getWeight() {
         return weight;
     }
